+++ dbt/include/databricks/macros/materializations/incremental/incremental.sql
@@ -38,6 +38,11 @@
     {%- endcall -%}
     {% do persist_constraints(target_relation, model) %}
     {% do apply_tags(target_relation, tags) %}
+    {%- if language == 'python' -%}
+      {%- do apply_tblproperties(target_relation, tblproperties) %}
+    {%- endif -%}
+
+    {% do persist_docs(target_relation, model, for_relation=language=='python') %}
   {%- elif existing_relation.is_view or existing_relation.is_materialized_view or existing_relation.is_streaming_table or should_full_refresh() -%}
     {#-- Relation must be dropped & recreated --#}
     {% set is_delta = (file_format == 'delta' and existing_relation.is_delta) %}
@@ -52,6 +57,7 @@
       {% do persist_constraints(target_relation, model) %}
     {% endif %}
     {% do apply_tags(target_relation, tags) %}
+    {% do persist_docs(target_relation, model, for_relation=language=='python') %}
   {%- else -%}
     {#-- Set Overwrite Mode to DYNAMIC for subsequent incremental operations --#}
     {%- if incremental_strategy == 'insert_overwrite' and partition_by -%}
@@ -94,17 +100,20 @@
     {%- endif -%}
     {% do apply_liquid_clustered_cols(target_relation) %}
     {% if _configuration_changes is not none %}
-      {% set tags = _configuration_changes.changes["tags"] %}
+      {% set tags = _configuration_changes.changes.get("tags", None) %}
+      {% set tblproperties = _configuration_changes.changes.get("tblproperties", None) %}
       {% if tags is not none %}
         {% do apply_tags(target_relation, tags.set_tags, tags.unset_tags) %}
       {%- endif -%}
+      {% if tblproperties is not none %}
+        {% do apply_tblproperties(target_relation, tblproperties.tblproperties) %}
+      {%- endif -%}
     {%- endif -%}
+    {% do persist_docs(target_relation, model, for_relation=True) %}
   {%- endif -%}
 
   {% set should_revoke = should_revoke(existing_relation, full_refresh_mode) %}
   {% do apply_grants(target_relation, grant_config, should_revoke) %}
-  {% do apply_tblproperties_python(target_relation, tblproperties, language) %}
-  {% do persist_docs(target_relation, model) %}
   {% do optimize(target_relation) %}
 
   {{ run_hooks(post_hooks) }}