+++ tests/unit/relation_configs/test_tblproperties.py
@@ -1,11 +1,12 @@
 import pytest
-from agate import Table
 from dbt.adapters.databricks.relation_configs.tblproperties import TblPropertiesConfig
 from dbt.adapters.databricks.relation_configs.tblproperties import (
     TblPropertiesProcessor,
 )
 from dbt.exceptions import DbtRuntimeError
 from mock import Mock
+
+from tests.unit import fixtures
 
 
 class TestTblPropertiesProcessor:
@@ -15,12 +16,14 @@ class TestTblPropertiesProcessor:
         assert spec == TblPropertiesConfig(tblproperties={})
 
     def test_from_results__single(self):
-        results = {"show_tblproperties": Table(rows=[["prop", "f1"]])}
+        results = {"show_tblproperties": fixtures.gen_tblproperties([["prop", "f1"]])}
         spec = TblPropertiesProcessor.from_relation_results(results)
         assert spec == TblPropertiesConfig(tblproperties={"prop": "f1"})
 
     def test_from_results__multiple(self):
-        results = {"show_tblproperties": Table(rows=[["prop", "1"], ["other", "other"]])}
+        results = {
+            "show_tblproperties": fixtures.gen_tblproperties([["prop", "1"], ["other", "other"]])
+        }
         spec = TblPropertiesProcessor.from_relation_results(results)
         assert spec == TblPropertiesConfig(tblproperties={"prop": "1", "other": "other"})